feat: cleanup and refactoring of constraints
