+++ packages/core/src/compiler/Style.ts
@@ -5,7 +5,8 @@ import {
   disambiguateSubNode,
 } from "compiler/Substance";
 import consola, { LogLevel } from "consola";
-import { constrDict, objDict } from "contrib/Constraints";
+import { constrDict } from "contrib/Constraints";
+import { objDict } from "contrib/Objectives";
 // Dicts (runtime data)
 import { compDict } from "contrib/Functions";
 import { constOf, numOf, varOf } from "engine/Autodiff";
