+++ packages/core/src/index.ts
@@ -346,7 +346,8 @@ export {
 export type { PenroseError } from "./types/errors";
 export * as Value from "./types/value";
 export type { Shape } from "./types/shape";
-export { objDict, constrDict } from "./contrib/Constraints";
+export { constrDict } from "./contrib/Constraints";
+export { objDict } from "./contrib/Objectives";
 export { compDict } from "./contrib/Functions";
 export type { Registry, Trio };
 export type { Env };
